# BitCaster.MatchingEngine.Contracts/specs/openapi.yaml
openapi: 3.1.0
info:
  title: BitCaster Matching Engine API
  description: REST API for the BitCaster prediction market matching engine.
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:5001
    description: Local development (via Aspire AppHost)
paths:
...
  /api/v1/{marketId}/orders:
    get:
      operationId: listRestingOrders
      summary: List the authenticated user's resting orders
      security:
        - nip98: []
      description: >
        Returns live resting orders in this market owned by the authenticated
        user. Used by clients that need to rebuild local order-tracking state
        after a restart.
      parameters:
        - $ref: '#/components/parameters/MarketId'
      responses:
        '200':
          description: Resting orders owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRestingOrdersResponse'
        '401':
          description: Missing or invalid NIP-98 authentication

    post:
      operationId: submitOrder
      summary: Submit a new order
      security:
        - nip98: []
      description: Submit a limit or market order to the matching engine.
      parameters:
        - $ref: '#/components/parameters/MarketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitOrderRequest'
      responses:
        '200':
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitOrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Missing or invalid NIP-98 authentication
  /api/v1/{marketId}/orders/{orderId}:
    get:
      operationId: getOrderStatus
      summary: Fetch the current state of an order
      security:
        - nip98: []
      description: >
        Returns the order's live status, remaining/filled amounts, and the
        list of fills produced so far. Clients poll this endpoint to observe
        async state changes (fills, cancellation, expiry) without holding an
        open connection. Only the order's owner (matching the authenticated
        pubkey) may query its status.
      parameters:
        - $ref: '#/components/parameters/MarketId'
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The order's unique identifier.
      responses:
        '200':
          description: Current order status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusResponse'
        '401':
          description: Missing or invalid NIP-98 authentication
        '404':
          description: Order not found

    delete:
      operationId: cancelOrder
      summary: Cancel an existing order
      security:
        - nip98: []
      parameters:
        - $ref: '#/components/parameters/MarketId'
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the order to cancel.
      responses:
        '200':
          description: Order cancelled
        '401':
          description: Missing or invalid NIP-98 authentication
        '404':
          description: Order not found
...
components:
  securitySchemes:
    nip98:
      type: http
      scheme: Nostr
      description: >
        NIP-98 HTTP Auth. The client signs a Nostr kind-27235 event containing
        the request URL and HTTP method, then sends it as
        `Authorization: Nostr <base64-encoded-event>`. The server verifies the
        BIP-340 Schnorr signature and extracts the signer's pubkey as the
        authenticated identity. See https://github.com/nostr-protocol/nips/blob/master/98.md
  parameters:
    MarketId:
      name: marketId
      in: path
      required: true
      schema:
        type: string
      description: >
        The market to trade on, in the format "{conditionId}-{outcomeSet}"
        (e.g. "deadbeef…abc-Alice" or "deadbeef…abc-B|C"). For finite
        categorical markets, multiple outcome names separated by "|" represent
        the order book for that exact outcome set. Clients should URL-encode
        the path segment when needed.
...
  schemas:
    Sats:
      type: integer
      format: int64
      minimum: 0
      description: A non-negative amount of satoshis.
    Probability:
      type: integer
      minimum: 1
      maximum: 99
      description: A probability price representing a percentage chance [1, 99].
    OrderSide:
      type: string
      enum:
        - Buy
        - Sell
      description: Direction of an order relative to the outcome token.
...
    TimeInForce:
      type: string
      enum:
        - GTC
        - FOK
        - FAK
      description: >
        Time-in-force policy. GTC = Good-Till-Cancel (rests on book),
        FOK = Fill-Or-Kill (reject if not fully filled),
        FAK = Fill-And-Kill (cancel remaining after partial fill).
    MatchPath:
      type: string
      enum:
        - Direct
        - Complementary
      description: How two orders were matched together.
    FillStatus:
      type: string
      enum:
        - Matched
        - Filled
        - Released
      description: >
        Lifecycle status of a fill-shaped order execution row. `Matched`
        means a DCB reservation exists and atomic-swap settlement is still
        pending; `Filled` means settlement committed; `Released` means a
        reservation failed or timed out and no longer consumes order depth.
...
    Fill:
      type: object
      required:
        - id
        - takerOrderId
        - makerOrderId
        - amountSats
        - executionPrice
        - path
        - filledAt
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique fill identifier.
        takerOrderId:
          type: string
          format: uuid
          description: The incoming (aggressor) order that triggered this fill.
        makerOrderId:
          type: string
          format: uuid
          description: The resting order that was matched against.
        amountSats:
          allOf:
            - $ref: '#/components/schemas/Sats'
          description: >
            Conditional-token face amount matched for settlement. DCB
            complementary reservations appear here before final fill commit so
            clients can join the atomic-swap TradeHub session.
        executionPrice:
          $ref: '#/components/schemas/Probability'
        path:
          $ref: '#/components/schemas/MatchPath'
        status:
          $ref: '#/components/schemas/FillStatus'
        filledAt:
          type: string
          format: date-time
          description: Timestamp when this fill was executed.
        tradeId:
          type: string
          format: uuid
          description: >
            Atomic-swap trade session identifier for this fill or DCB
            reservation. Present when the client must join TradeHub to settle;
            omitted only for legacy fills that do not have a corresponding
            TradeHub session.
        makerEphemeralPubkey:
          type: string
          pattern: '^[0-9a-fA-F]{66}$'
          description: >
            Hex-encoded compressed secp256k1 pubkey of the maker order's
            ephemeral key. Present on direct-match fills so the taker can
            derive the ECDH shared secret with the maker without an extra
            round-trip through the engine. Null on complementary-match fills
            and on fills against orders that did not declare an ephemeral
            pubkey (e.g. legacy automated-liquidity orders).
    SubmitOrderRequest:
      type: object
      required:
        - outcomeId
        - side
        - price
        - amountSats
        - ephemeralPubkey
      properties:
        outcomeId:
          type: string
          description: >
            The outcome or finite outcome set to trade (e.g. "Alice", "YES",
            or "B|C"). Must match the outcome-set segment of marketId.
        side:
          $ref: '#/components/schemas/OrderSide'
        price:
          $ref: '#/components/schemas/Probability'
        amountSats:
          allOf:
            - $ref: '#/components/schemas/Sats'
          description: >
            Limit-order size as conditional-token face amount. First release
            requires this to be divisible by 100 sats so integer-cent prices
            produce exact quote payments.
        timeInForce:
          $ref: '#/components/schemas/TimeInForce'
          default: GTC
        ephemeralPubkey:
          type: string
          pattern: '^[0-9a-fA-F]{66}$'
          description: >
            Hex-encoded compressed secp256k1 pubkey (33 bytes, 66 hex chars)
            generated by the client per-order. Used by the matching engine to
            route the counterparty's encrypted atomic-swap messages back to
            this order's submitter. The corresponding private key stays on the
            client and is the only thing that can decrypt and sign swap
            payloads — the engine never sees it.
    OrderStatusResponse:
      type: object
      required:
        - orderId
        - marketId
        - status
        - remainingAmountSats
        - filledAmountSats
        - fills
      properties:
        orderId:
          type: string
          format: uuid
          description: The order's unique identifier.
        marketId:
          type: string
          description: The market this order belongs to.
        status:
          type: string
          description: >
            One of: "resting" (on book, unmatched), "matched" (reserved for
            atomic-swap settlement), "partially_filled", "filled",
            "cancelled".
        remainingAmountSats:
          $ref: '#/components/schemas/Sats'
        filledAmountSats:
          $ref: '#/components/schemas/Sats'
          description: >
            Conditional-token face amount already consumed by fills or active
            DCB complementary reservations. A reservation is exposed here
            before final settlement so clients can notify makers and start
            the atomic-swap handshake.
        fills:
          type: array
          items:
            $ref: '#/components/schemas/Fill'
          description: >
            All fills and active DCB complementary reservation handles
            produced against this order so far.
    RestingOrderResponse:
      type: object
      required:
        - orderId
        - marketId
        - outcomeId
        - side
        - price
        - remainingAmountSats
        - amountSats
        - timeInForce
        - placedAt
      properties:
        orderId:
          type: string
          format: uuid
          description: The unique identifier assigned by the matching engine.
        marketId:
          type: string
          description: The market this order belongs to.
        outcomeId:
          type: string
          description: The outcome this order trades.
        side:
          $ref: '#/components/schemas/OrderSide'
        price:
          $ref: '#/components/schemas/Probability'
        remainingAmountSats:
          $ref: '#/components/schemas/Sats'
        amountSats:
          $ref: '#/components/schemas/Sats'
        timeInForce:
          $ref: '#/components/schemas/TimeInForce'
        placedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        ephemeralPubkey:
          type: string
          nullable: true
          pattern: '^[0-9a-fA-F]{66}$'
          description: Order-level ephemeral pubkey supplied at submit time.
    ListRestingOrdersResponse:
      type: object
      required:
        - orders
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/RestingOrderResponse'
    SubmitOrderResponse:
      type: object
      required:
        - orderId
        - status
        - remainingAmountSats
        - fills
        - ephemeralPubkey
      properties:
        orderId:
          type: string
          format: uuid
          description: The unique identifier assigned to this order.
        status:
          type: string
          description: >
            One of: "filled", "matched" (reserved for atomic-swap
            settlement), "partially_filled", "resting", "cancelled".
        remainingAmountSats:
          $ref: '#/components/schemas/Sats'
        fills:
          type: array
          items:
            $ref: '#/components/schemas/Fill'
          description: List of fills produced by this order. Empty if no matches.
        ephemeralPubkey:
          type: string
          pattern: '^[0-9a-fA-F]{66}$'
          description: >
            Echoes the ephemeralPubkey the client submitted. Lets clients
            confirm round-trip before they discard the corresponding privkey.
...